Firstparty converted to proper static lib + utils gutted and split into firstparty + widgets moved to firstparty + build flags fixes and reordering - feat: `Firstparty` is now a proper module compiled down to a static lib. Note: The issue is that for any needed STB-style lib, two include are needed: one for the proper header, the other mainly at the end, with the implementation macro. That was becoming increasingly annoying with the more header needed, so they were converted into distinct header and src part and compile down to single static lib. - fix: Added `f32Vector2_def` guard to prevent macro redefinition hell to happen. - feat: Gutted `utils` and split into distinct headers and sources into `firstparty`. The reason boils down to: `utils.c`/`h` was at first duplicated in both `lobby` and `bingo`, that was enough to put it into `firstparty` after having thought over it for many hours. - chore: Removed gutted utils off bingo and lobby. - build/fix: Added back `BASE_CFLAGS` and `BASE_LDFLAGS` that was removed I don't where/when. - build: Renamed `CFLAG` and `LDFLAGS` specific to mode to respectively `MODE_CFLAGS` and `MODE_LDFLAGS` - build/fix: Reorder when each CFLAGS/LDFLAGS appeared for compilation/linkage purposes. - refactor: Applied new `utils` split up to places where `utils/utils.h` was included. - style: Continued to apply CSC whenever I see the needs for it. - refactor/feat: Moved `widgets` folder into `firstparty` to be reuseable everywhere in the project

+++ firstparty/src/widgets/slider.c
@@ -12,12 +12,13 @@
             - Optional value label
 */
 
+#include <stdio.h>
+#include <math.h>
+
 #include "widgets/slider.h"
 #include "widgets/types.h"
 
-#include "utils/utils.h"
-
-#include "systemSettings.h"
+#include "utils/mathUtils.h"
 
 // ─────────────────────────────────────────────────────────────────────────────
 
@@ -46,7 +47,7 @@ bool sliderUpdate(Slider_St* slider, Vector2 mouseScreen) {
         f32 oldValue = slider->value;
         slider->value = slider->minValue + t * (slider->maxValue - slider->minValue);
 
-        return (fabsf(slider->value - oldValue) > 0.0001f);
+        return (fabs(slider->value - oldValue) > 0.0001f);
     }
 
     if (IsMouseButtonReleased(MOUSE_BUTTON_LEFT)) {
